+++ lib/src/ui/pages/painting/widgets.dart
@@ -1,9 +1,17 @@
 import 'package:dreambook/src/ui/pages/shared/shared_code_view.dart';
 
 import 'backdrop_filter.dart';
+import 'clip.dart';
 import 'decorated_box.dart';
+import 'opacity.dart';
+import 'rotated_box.dart';
+import 'transform.dart';
 
 final paintingCodeItems = <CodeItem>[
   backdropFilterBoxItem,
   decoratedBoxItem,
+  opacityItem,
+  clipItem,
+  rotatedBoxItem,
+  transformItem,
 ];