+++ lib/src/ui/theme.g.dart
@@ -6,7 +6,7 @@ part of 'theme.dart';
 // RiverpodGenerator
 // **************************************************************************
 
-String _$themeModeImageHash() => r'59b0696254faf39c1f9569e91cd28ce31cd6b8c1';
+String _$themeModeImageHash() => r'e42010bc5769de4669eae28d53f28d63f2b4fc9e';
 
 /// See also [themeModeImage].
 @ProviderFor(themeModeImage)