+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+### Added
+- add transform,rotatedBox,clip,opacity widgets
+
 ## 1.0.1+2 - 2023-10-27
 ### Added
 - add github button